fix: sentry calls
Refactor: envs configs and sentry utils
