+++ package.json
@@ -2,7 +2,7 @@
 	"$schema": "https://json.schemastore.org/package.json",
 	"name": "node_backend_boilerplate",
 	"description": "Node.js Boilerplate for Back-End using TypeScript",
-	"version": "4.15.1",
+	"version": "4.15.2",
 	"license": "MIT",
 	"author": "Álvaro Davi Santos Alves <alvaro.davisa@gmail.com>",
 	"contributors": [
@@ -21,8 +21,9 @@
 		"win32"
 	],
 	"engines": {
-		"node": "f>=18.6.x <=20.x.x",
-		"yarn": "f>=1.22.x <3.x.x"
+		"node": "f>=18.6.x <=22.x.x",
+		"npm": "f>=10.9.x <=11.x.x",
+		"yarn": "f>=1.22.x <2.x.x"
 	},
 	"engineStrict": true,
 	"keywords": [
