+++ src/modules/common/filters/AbstractExceptions.filter.ts
@@ -1,12 +1,12 @@
 import { HttpException } from '@nestjs/common';
 import { WsException } from '@nestjs/websockets';
 import { AxiosError } from 'axios';
-import { captureException } from '@core/errors/trackers';
 import LoggerService from '@core/logging/Logger.service';
 import DataParserHelper from '@common/utils/helpers/DataParser.helper';
 import { HttpStatusEnum } from '@common/enums/httpStatus.enum';
 import { ExceptionsEnum } from '@common/enums/exceptions.enum';
 import { getObjValues } from '@common/utils/dataValidations.util';
+import { captureException } from '@common/utils/sentryCalls.util';
 import { ExceptionMetadataInterface } from '@shared/internal/interfaces/errorInterface';
