+++ src/modules/core/configs/swaggerDoc.config.ts
@@ -1,16 +1,17 @@
 import { INestApplication } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import readPackageInfo from '@common/utils/packageInfoReader.util';
+import envsConfig from './envs.config';
 
+
+const { package: { version: packageVersion, description: packageDescription } } = envsConfig();
 
 export default (nestApp: INestApplication): void => {
-	const { description, version } = readPackageInfo();
-	const apiDescription = description ? `${description} - API` : undefined;
+	const apiDescription = packageDescription ? `${packageDescription} - API` : undefined;
 
 	const config = new DocumentBuilder()
 		.setTitle('Node Back-End Boilerplate')
 		.setDescription(apiDescription ?? 'API Boilerplate created with Nest.js')
-		.setVersion(version ?? '1.0.0')
+		.setVersion(packageVersion ?? '1.0.0')
 		.setContact('Álvaro Alves', 'https://github.com/AlvaroDavi5', 'alvaro.davisa@gmail.com')
 		.addServer('http://localhost:3000', 'Main Server', {})
 		.addServer('http://localhost:4000', 'Mocked Server', {})
