+++ src/modules/core/core.module.ts
@@ -10,6 +10,7 @@ import { SentryModule } from '@sentry/nestjs/setup';
 import { DevtoolsModule } from '@nestjs/devtools-integration';
 import AppModule from '@app/app.module';
 import GraphQlModule from '@graphql/graphql.module';
+import { formatGraphQlError } from '@graphql/utils/errors.util';
 import EventsModule from '@events/events.module';
 import RequestRateLimitConstants from '@common/constants/RequestRateLimit.constants';
 import { EnvironmentsEnum } from '@common/enums/environments.enum';
@@ -17,7 +18,6 @@ import CommonModule from '@common/common.module';
 import envsConfig from './configs/envs.config';
 import LifecycleService from './start/Lifecycle.service';
 import Exceptions from './errors/Exceptions';
-import { formatGraphQlError } from './errors/trackers';
 import LoggerService, { RequestLoggerProvider } from './logging/Logger.service';
 import CryptographyService from './security/Cryptography.service';
 import DatabaseConnectionProvider from './infra/database/connection';
