+++ src/modules/core/logging/Logger.service.ts
@@ -3,8 +3,8 @@ import { Injectable, Inject, Provider, Scope } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { createLogger, Logger } from 'winston';
 import { ConfigsInterface } from '@core/configs/envs.config';
-import { captureLog } from '@core/errors/trackers';
 import DataParserHelper from '@common/utils/helpers/DataParser.helper';
+import { captureLog } from '@common/utils/sentryCalls.util';
 import { LoggerInterface, LogLevelEnum, MetadataInterface, getLoggerOptions } from './logger';
